+++ packages/metascraper-clearbit/index.js
@@ -1,15 +1,12 @@
 'use strict'
 
+const { memoizeOne, composeRule } = require('@metascraper/helpers')
 const { get, isString, isObject } = require('lodash')
-const { createValidator } = require('@metascraper/helpers')
 const { stringify } = require('querystring')
-const memoizeOne = require('memoize-one')
 const { getDomain } = require('tldts')
 const got = require('got')
 
 const ENDPOINT = 'https://autocomplete.clearbit.com/v1/companies/suggest'
-
-const memoFn = (newArgs, oldArgs) => newArgs[0].url === oldArgs[0].url
 
 const DEFAULT_GOT_OPTS = {
   json: true,
@@ -25,7 +22,7 @@ const appendQuery = (data, query) => {
 }
 
 const createClearbit = ({ gotOpts, logoOpts } = {}) =>
-  memoizeOne(async ({ url }) => {
+  memoizeOne(async ($, url) => {
     const domain = getDomain(url)
 
     try {
@@ -38,10 +35,10 @@ const createClearbit = ({ gotOpts, logoOpts } = {}) =>
     } catch (err) {
       return null
     }
-  }, memoFn)
+  })
 
 module.exports = opts => {
-  const getClearbit = createValidator(createClearbit(opts))
+  const getClearbit = composeRule(createClearbit(opts))
 
   return {
     logo: getClearbit({ from: 'logo' }),