+++ packages/metascraper-helpers/index.js
@@ -1,12 +1,15 @@
 'use strict'
 
 const {
+  castArray,
   chain,
   eq,
   flow,
   get,
   includes,
+  invoke,
   isArray,
+  isDate,
   isEmpty,
   isNumber,
   isString,
@@ -14,10 +17,7 @@ const {
   replace,
   size,
   toLower,
-  trim,
-  invoke,
-  isNil,
-  castArray
+  trim
 } = require('lodash')
 
 const langs = require('iso-639-3').map(({ iso6391 }) => iso6391)
@@ -29,6 +29,7 @@ const fileExtension = require('file-extension')
 const _normalizeUrl = require('normalize-url')
 const smartquotes = require('smartquotes')
 const { decodeHTML } = require('entities')
+const memoizeOne = require('memoize-one')
 const mimeTypes = require('mime-types')
 const hasValues = require('has-values')
 const chrono = require('chrono-node')
@@ -37,7 +38,6 @@ const isIso = require('isostring')
 const toTitle = require('title')
 const isUri = require('is-uri')
 const { URL } = require('url')
-const memoizeOne = require('memoize-one')
 
 const VIDEO = 'video'
 const AUDIO = 'audio'
@@ -131,11 +131,11 @@ const protocol = url => {
   return protocol.replace(':', '')
 }
 
-const isMediaTypeUrl = (url, type, opts) =>
-  isUrl(url, opts) && isMediaTypeExtension(url, type)
+const isMediaTypeUrl = (url, type, { ext, ...opts } = {}) =>
+  isUrl(url, opts) && isMediaTypeExtension(url, type, ext)
 
-const isMediaTypeExtension = (url, type) =>
-  eq(type, get(EXTENSIONS, extension(url)))
+const isMediaTypeExtension = (url, type, ext) =>
+  eq(type, get(EXTENSIONS, ext || extension(url)))
 
 const isMediaUrl = (url, opts) =>
   isImageUrl(url, opts) || isVideoUrl(url, opts) || isAudioUrl(url, opts)
@@ -203,7 +203,7 @@ const date = value => {
 
   // try to parse a complex date string
   const parsed = chrono.parseDate(value)
-  if (parsed) return parsed.toISOString()
+  if (isDate(parsed)) return parsed.toISOString()
 }
 
 const lang = value => {
@@ -253,12 +253,12 @@ const logo = url
 
 const video = (value, opts) => {
   const urlValue = url(value, opts)
-  return isVideoUrl(urlValue) && urlValue
+  return isVideoUrl(urlValue, opts) && urlValue
 }
 
 const audio = (value, opts) => {
   const urlValue = url(value, opts)
-  return isAudioUrl(urlValue) && urlValue
+  return isAudioUrl(urlValue, opts) && urlValue
 }
 
 const validator = {
@@ -275,27 +275,22 @@ const validator = {
   lang
 }
 
-/**
- * Create a property mapper with validator inside.
- */
-const createValidator = fn => ({ from, to = from }) => async args => {
-  const data = await fn(args)
-  const value = get(data, from)
-  return invoke(validator, to, value)
-}
-
-/**
- * Wrap a rule into a validator
- */
-const createWrap = (fn, opts) => rule => ({ htmlDom, url }) => {
+const toRule = (fn, opts) => rule => ({ htmlDom, url }) => {
   const value = rule(htmlDom, url)
-  return fn(value, opts)
+  return fn(value, { url, ...opts })
 }
 
-const hasValue = value =>
-  isNil(value) || value === false || value === 0 || value === ''
-    ? false
-    : hasValues(value)
+const composeRule = fn => ({ from, to = from, ...opts }) => async ({
+  htmlDom,
+  url
+}) => {
+  const data = await fn(htmlDom, url)
+  const value = get(data, from)
+  return invoke(validator, to, value, { url, ...opts })
+}
+
+const has = value =>
+  value === null || value === false || value === 0 ? false : hasValues(value)
 
 module.exports = {
   $filter,
@@ -304,12 +299,11 @@ module.exports = {
   audio,
   audioExtensions,
   author,
-  createValidator,
-  createWrap,
+  composeRule,
   date,
   description,
   extension,
-  hasValue,
+  has,
   image,
   imageExtensions,
   isArray,
@@ -328,6 +322,7 @@ module.exports = {
   jsonld,
   lang,
   logo,
+  memoizeOne,
   normalizeUrl,
   protocol,
   publisher,
@@ -337,5 +332,6 @@ module.exports = {
   url,
   validator,
   video,
-  videoExtensions
+  videoExtensions,
+  toRule
 }