+++ packages/metascraper-readability/index.js
@@ -1,23 +1,21 @@
 'use strict'
 
-const { createValidator } = require('@metascraper/helpers')
+const { memoizeOne, composeRule } = require('@metascraper/helpers')
+
 const Readability = require('readability')
-const memoizeOne = require('memoize-one')
 const jsdom = require('jsdom')
 
 const { JSDOM } = jsdom
 
-const memoFn = (newArgs, oldArgs) => newArgs[0].url === oldArgs[0].url
-
 const virtualConsole = new jsdom.VirtualConsole()
 
-const readability = memoizeOne(({ htmlDom, url }) => {
-  const dom = new JSDOM(htmlDom.html(), { virtualConsole, url })
+const readability = memoizeOne(($, url) => {
+  const dom = new JSDOM($.html(), { virtualConsole, url })
   const reader = new Readability(dom.window.document)
   return reader.parse()
-}, memoFn)
+})
 
-const getReadbility = createValidator(readability)
+const getReadbility = composeRule(readability)
 
 module.exports = () => {
   return {