+++ packages/metascraper-uol/index.js
@@ -4,31 +4,32 @@ const {
   $jsonld,
   title,
   description,
-  createWrap
+  toRule,
+  memoizeOne
 } = require('@metascraper/helpers')
-const memoizeOne = require('memoize-one')
 const { getDomain } = require('tldts')
 
 const ROOT_DOMAINS = ['uol.com.br', 'torcedores.com']
 
-const isValidUrl = memoizeOne(url =>
+const isValidUrl = memoizeOne(({ url }) =>
   ROOT_DOMAINS.some(domain => getDomain(url) === domain)
 )
 
-const wrapTitle = createWrap(title)
-const wrapDescription = createWrap(description)
+const toTitle = toRule(title)
+const toDescription = toRule(description)
 
 module.exports = () => {
   const rules = {
     title: [
-      wrapTitle(($, url) => $jsonld('headline')($, url)),
-      wrapTitle(($, url) => $jsonld('name')($, url)),
-      wrapTitle($ => $('title').text())
+      toTitle(($, url) => $jsonld('headline')($, url)),
+      toTitle(($, url) => $jsonld('name')($, url)),
+      toTitle($ => $('title').text())
     ],
-    description: [wrapDescription(($, url) => $jsonld('description')($, url))]
+    description: [toDescription(($, url) => $jsonld('description')($, url))]
   }
 
-  rules.test = ({ url }) => isValidUrl(url)
+  rules.test = isValidUrl
+
   return rules
 }