+++ packages/metascraper/src/get-data.js
@@ -1,11 +1,11 @@
 'use strict'
 
 const { isString, map, fromPairs } = require('lodash')
-const { hasValue } = require('@metascraper/helpers')
+const { has } = require('@metascraper/helpers')
 const mapValuesDeep = require('map-values-deep')
 const xss = require('xss')
 
-const noopTest = () => true
+const truthyTest = () => true
 
 const getValue = async ({ htmlDom, url, rules, meta, ...props }) => {
   const lastIndex = rules.length
@@ -14,17 +14,18 @@ const getValue = async ({ htmlDom, url, rules, meta, ...props }) => {
 
   do {
     const rule = rules[index++]
-    const test = rule.test || noopTest
+    const test = rule.test || truthyTest
+
     if (test({ htmlDom, url, meta })) {
       value = await rule({ htmlDom, url, meta, ...props })
     }
-  } while (!hasValue(value) && index < lastIndex)
+  } while (!has(value) && index < lastIndex)
 
   return value
 }
 
 const escapeValue = (value, { escape }) => {
-  if (!hasValue(value)) return null
+  if (!has(value)) return null
   if (!escape) return value
   return mapValuesDeep(value, value => (isString(value) ? xss(value) : value))
 }