+++ CHANGELOG.md
@@ -3,6 +3,82 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.8.6](https://github.com/microlinkhq/metascraper/compare/v5.8.5...v5.8.6) (2019-11-11)
+
+**Note:** Version bump only for package metascraper
+
+
+
+
+
+## [5.8.5](https://github.com/microlinkhq/metascraper/compare/v5.8.4...v5.8.5) (2019-11-11)
+
+
+### Bug Fixes
+
+* ensure chrono output is a date ([2391f15](https://github.com/microlinkhq/metascraper/commit/2391f15a31de63943a4296baaec21b632da0a5c0))
+
+
+
+
+
+## [5.8.4](https://github.com/microlinkhq/metascraper/compare/v5.8.3...v5.8.4) (2019-11-11)
+
+**Note:** Version bump only for package metascraper
+
+
+
+
+
+## [5.8.3](https://github.com/microlinkhq/metascraper/compare/v5.8.2...v5.8.3) (2019-11-09)
+
+
+### Bug Fixes
+
+* ensure only spotify urls are valid ([a707c57](https://github.com/microlinkhq/metascraper/commit/a707c574d1afd0f1765260bba69b7fbb29a90600))
+
+
+
+
+
+## [5.8.2](https://github.com/microlinkhq/metascraper/compare/v5.8.1...v5.8.2) (2019-11-09)
+
+**Note:** Version bump only for package metascraper
+
+
+
+
+
+## [5.8.1](https://github.com/microlinkhq/metascraper/compare/v5.8.0...v5.8.1) (2019-11-09)
+
+
+### Bug Fixes
+
+* ensure innerRules is a collection ([#234](https://github.com/microlinkhq/metascraper/issues/234)) ([04fe20d](https://github.com/microlinkhq/metascraper/commit/04fe20dd9db5686d778ad614f69f545728368dbb))
+
+
+
+
+
+# [5.8.0](https://github.com/microlinkhq/metascraper/compare/v5.7.21...v5.8.0) (2019-11-08)
+
+
+### Features
+
+* add spotify integration ([#232](https://github.com/microlinkhq/metascraper/issues/232)) ([74cd3ac](https://github.com/microlinkhq/metascraper/commit/74cd3ac112323baef229b5b49075343997bc628a))
+
+
+
+
+
+## [5.7.21](https://github.com/microlinkhq/metascraper/compare/v5.7.20...v5.7.21) (2019-11-08)
+
+**Note:** Version bump only for package metascraper
+
+
+
+
+
 ## [5.7.20](https://github.com/microlinkhq/metascraper/compare/v5.7.19...v5.7.20) (2019-11-07)
 
 **Note:** Version bump only for package metascraper