+++ packages/metascraper-video/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.8.6](https://github.com/microlinkhq/metascraper/tree/master/packages/metascraper-video/compare/v5.8.5...v5.8.6) (2019-11-11)
+
+**Note:** Version bump only for package metascraper-video
+
+
+
+
+
+## [5.8.5](https://github.com/microlinkhq/metascraper/tree/master/packages/metascraper-video/compare/v5.8.4...v5.8.5) (2019-11-11)
+
+**Note:** Version bump only for package metascraper-video
+
+
+
+
+
+# [5.8.0](https://github.com/microlinkhq/metascraper/tree/master/packages/metascraper-video/compare/v5.7.21...v5.8.0) (2019-11-08)
+
+**Note:** Version bump only for package metascraper-video
+
+
+
+
+
+## [5.7.21](https://github.com/microlinkhq/metascraper/tree/master/packages/metascraper-video/compare/v5.7.20...v5.7.21) (2019-11-08)
+
+**Note:** Version bump only for package metascraper-video
+
+
+
+
+
 ## [5.7.17](https://github.com/microlinkhq/metascraper/tree/master/packages/metascraper-video/compare/v5.7.16...v5.7.17) (2019-10-29)
 
 **Note:** Version bump only for package metascraper-video